Share one UninitBuf between bun_sys, the runtime and the TLS wrapper
Lives in bun_core::vec next to the spare-capacity helpers it mirrors, so
the uws wrapper can use it too instead of its own MaybeUninit wrapper.
